+++ Demo/KeyboardPro/KeyboardViewController.swift
@@ -62,7 +62,7 @@ class KeyboardViewController: KeyboardInputViewController {
         ///
         /// The default haptic feedback is `.minimal`, which
         /// only has haptic feedback for long press on space.
-        state.feedbackConfiguration.audioConfiguration.delete = .custom(id: 1329)
+        // state.feedbackConfiguration.audioConfiguration.delete = .custom(id: 1329)
         state.feedbackConfiguration.isHapticFeedbackEnabled = true
 
         /// 💡 Call super to perform the base initialization.