+++ Sources/KeyboardKit/_Deprecations/AutocompleteContext+Deprecated.swift
@@ -3,6 +3,7 @@
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2023-12-06.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
 //
 
 import Foundation