+++ Sources/KeyboardKit/_Keyboard/KeyboardInputViewController.swift
@@ -31,8 +31,7 @@ open class KeyboardInputViewController: UIInputViewController, KeyboardControlle
 
     open override func viewDidLoad() {
         super.viewDidLoad()
-        setupInitialWidth()
-        setupLocaleObservation()
+        setupController()
         viewWillRegisterSharedController()
     }
 
@@ -125,15 +124,6 @@ open class KeyboardInputViewController: UIInputViewController, KeyboardControlle
         setup(withRootView: Keyboard.RootView { [unowned self] in
             view(self)
         })
-    }
-
-    /// Setup KeyboardKit with a custom root view.
-    func setup<Content: View>(withRootView view: Content) {
-        self.children.forEach { $0.removeFromParent() }
-        self.view.subviews.forEach { $0.removeFromSuperview() }
-        let view = view.withEnvironment(fromController: self)
-        let host = KeyboardHostingController(rootView: view)
-        host.add(to: self)
     }
 
 
@@ -325,21 +315,6 @@ open class KeyboardInputViewController: UIInputViewController, KeyboardControlle
 // MARK: - Private Functions
 
 private extension KeyboardInputViewController {
-
-    /// Set up an initial width to avoid SwiftUI layout bugs.
-    func setupInitialWidth() {
-        view.frame.size.width = UIScreen.main.bounds.width
-    }
-
-    /// Setup locale observation to handle locale changes.
-    func setupLocaleObservation() {
-        state.keyboardContext.$locale.sink { [weak self] in
-            guard let self = self else { return }
-            let locale = $0
-            self.primaryLanguage = locale.identifier
-            self.services.autocompleteProvider.locale = locale
-        }.store(in: &cancellables)
-    }
 
     func tryChangeToPreferredKeyboardTypeAfterTextDidChange() {
         let shouldSwitch = services.keyboardBehavior.shouldSwitchToPreferredKeyboardTypeAfterTextDidChange()